Remove Red Hat internal references from Dockerfiles
This commit makes the existing Dockerfiles and docker-compose files a
bit more modular so that the Red Hat internal components can be
extracted and provided as part of the infrastructure setup instead of
being hardcoded into the application itself.

This has the benefit of separation of concerns and making OSIDB more
accessible to third parties.