Allow unauthenticated access to the OSIDB API
This commit changes the permission class of the main OSIDB endpoints to
IsAuthenticatedOrReadOnly which allows unauthenticated users to consume
the API in read-only fashion.

The commit also applies some changes to the auth and middleware hooks
that set the ACLs so that AnonymousUsers (=unauthenticated users) can
only read publically available information.

Closes OSIDB-313