Get all affects to sync from existing trackers
Before this commit, the function of the Jiraffe collector that would
fetch all trackers and their affects in order to calculate which
trackers to sync would access the `affect` attribute on `Tracker`,
however this would yield an `AttributeError` due to the fact that the
Tracker<->Affect relationship is ManyToMany and the field name is thus
`affects` and not `affect`.

This commit corrects the error by accessing the `affects` field instead
and iterating over each affect.