Remove pip-audit checks
After a brief discussion we informally decided to drop pip-audit usage,
the main argument being that whenever a vulnerability was discovered in
one of our packages our entire development process is halted -- CI
cannot pass as long as the pip-audit check is failing, so we must
immediately update to the latest version in order to be able to continue
normal development and often times the vulnerabilities don't even affect
us since we may not use the feature that the security patch fixes.

Since we develop on GitHub now, we can simply defer this job to
Dependabot which can alert us once a week (by default) of potential
security vulnerabilities present in our codebase and let developers
gauge whether it's urgent or not to apply the security patches, without
disrupting workflow.

For reference, Dependabot is highly customizable through a
dependabot.yaml file inside the .github directory of the repository:
https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file