Prevent pip-compile from adding the index URL to requirements files
Since we allow the usage of a custom pip index, running the pip-compile
command as it was would add the --index-url flag to the resulting
requirements.txt or devel-requirements.txt file which might be unwanted
if the pip repository is private.

With this commit we add the --no-emit-index-url flag to the pip-compile
command which prevents it from adding the --index-url argument to the
resulting requirements files.