Implement mixin for the creation of alerts in relevant models
This commit introduces a mixin called AlertMixin which provides any
inheriting model with features for creating alerts -- these alerts are
stored in a JSONField and are particular to each record of a given
model, it is also a scalar field meaning that it's easy to serialize and
performant (no table joins necessary).

Once a model inherits from the mixin, a helper method `alert()` becomes
available on instances of the model, which is the interface through
which alerts can be created as it guarantees a certain level of schema
consistency.

Closes OSIDB-324