Make _alerts field a non-required field when using AlertMixin
Having the _alerts field without blank=True means that Django will
require it when creating objects of the models that inherit from said
mixin, despite the field definition having a default set.

This meant that creation of objects without explicitly providing the
_alerts field would trigger an error on save.

This commit makes the field blank=True so that this is not a blocker as
_alerts is never expected to be supplied by hand, and should only be
triggered by "the system" during validation.