Make Affect model inherit from AlertMixin
The migration for this change has been written by hand so as to prevent
any potential performance bottlenecks from happening when it is applied
on production.

This migration should keep the PostgreSQL default intact (which Django
usually strips and does Python-side only) so that there's no need to
rewrite the entire table and/or populate every record from Affect.