Check that an Affect's ps_module is valid on save
This commit implements two validations that will check whether the
ps_module for a given Affect record is valid when saving the record.

* The one will only impact "legacy" flaws, this is determined based on
  a sentinel bugzilla ticket id, any flaw whose bugzilla ticket id
  matches exactly the sentinel or is lower than it is considered a
  "legacy" flaw, for said flaws the validation will simply add an alert
  by leveraging the AlertMixin inheritance in the Affect model.

* The other will impact any "newer" flaws that were created after the
  sentinel bugzilla ticket id and any truly new flaws, for which a new
  convention was used of using ps_module and not ps_update_streams, in
  this case a ValidationError will be raised.

In the future we might fix the data via some sort of heuristic and
finish it off with some manual work, at that point we might only keep
the latter of the two validations to have a more robust validation.