Prevent libpq from attempting GSSAPI authentication
As it turns out, postgresql 12 introduced support for secure GSS
communication between libpq and pg over TCP/IP if libpq is compiled with
GSSAPI support through an option called gssencmode. This option, by
default, is set to "prefer", meaning that if libpq detects GSSAPI
credentials in the system it's running on, it will attempt to connect to
the database via encrypted GSSAPI connection.

Since we do not have GSSAPI connection properly set up between the OSIDB
service pod and the database pod, this will fail, always, leading to
wasted resources on every request to the database.

https://www.postgresql.org/docs/13/libpq-connect.html#LIBPQ-CONNECT-GSSENCMODE

Closes OSIDB-424

Co-authored-by: Martin Prpic <mprpic@redhat.com>