Handle FlawSource that are both public and private differently
A few flaw sources such as GENTOO, SUSE and UBUNTU can be both public
and private.

This commit changes the way existing and future validations based on
sources work, by providing an extra `is_private` method on FlawSource we
can now filter sources that are both public and private, and in those
cases the better course of action is probably to create an alert and not
raise any ValidationErrors.

Closes OSIDB-450