Add various env vars to osidb-service in compose file
Usually these are not necessary to exist inside the actual web service
because most of the time the web service does not contact
ET/DASHBOARD/PRODUCT_DEFS directly, instead that is done through celery
workers.

However in order to be able to have a custom Django management command
that syncs flaws these variables must exist within the web service pod,
which is why they have been added in this commit.