Fix update_or_create usage in Errata linking
et_id is unique but advisory_name may change and when the change occurs
the collector tried to create a new Erratum with the same et_id and
new advisory_name which resulted in integrity error (duplicate et_id).
This change alters the update_or_create usage so the Erratum is searched
just by its et_id and the advisory_name is updated accordingly.