Give the osidb.acl parameter a default value from migrations
Running the initial migrations on a brand new managed database will
result in an error when setting up row-level security because the SQL
snippet that sets it up expects there to be an osidb.acl parameter
defined in the settings.

Since we may not always have access to the postgresql.conf file or we
might not want to override an existing one, we can simply set the
parameter to a default value before applying the row-level security
checks.

This change should not impact existing databases either.