Add check for CHANGELOG.md on all Pull Requests
This commit adds the very first GitHub workflow to the OSIDB repo and
the very first action which checks that every PR updates the
CHANGELOG.md, if the PR touches on changes that do not require a
CHANGELOG entry (e.g. purely technical changes) the `technical` label
can be added to the PR to bypass the check.

The commit also updates the documentation to reflect these changes and
moves stuff around a bit so that it makes more sense.