Ignore OSIM tests by default when running testrunner entrypoints
This commit makes it so that OSIM tests are ignored for every single
make/tox entrypoint except for the one that specifically tests OSIM
(testrunner.osim).

This commit also ensures that both the local and ci tests being run are
as similar as possible (both exclude OSIM, CI excludes integration tests).