Remove unused image from dev environment
Phpldapadmin is supposed to be a configuration UI for the ldap accounts
that are used in testing, but since they're all set up already there's
no reason to use this image.

Even if we did need to manage the ldap accounts, changing the ldif
should be easy enough to do, a webUI is not required.

Furthermore, using the latest tag of images like these is a potential
supply chain vulnerability waiting to happen.

Closes OSIDB-639