Ensure pytest captures stdout/stderr logs when running tests
As long as the test passes the logs shouldn't be an issue, by default
pytest will still show logs from tests that fail so this should be ok.

This commit also removes the warning filters as there are some real
warnings being skipped and they should be taken more seriously as they
might lead to real issues.