Make first login into OSIDB fault-tolerant
On a user's first login we trigger a signal that creates a new Profile
object which will link the kerberos login to Bugzilla and JIRA
usernames, however this can fail for a number of reasons (expired token,
service unavailable, network failure, etc.) and thus we should take care
to handle such exceptions and simply let the user login without linking
said credentials.

In the future we should try to sync at other points in time if possible.