Disable venv checks on CI
This commit adds a check to the `check-venv` and `check-venv-active`
entrypoints that will bypass said checks if running in a CI (CI env
variable is not null), this is because the CI infrastructure does not
use (nor need) virtualenvs.

While the scripts still run successfully in CI regardless of these
checks, the error messages in the log can be confusing and/or
distracting.