Don't rely on package cache for GH actions
The default package cache in the GH actions ubuntu-latest image includes
an URI for the libsasl2-dev package that is no longer being served, thus
when installing the package it fails and the checks that depend on this
package fail.

To fix it, we first do a package database update and then we install the
packages.