Make embargoed a custom field
Embargoed field is not really a field in OSIDB, it just modifies the
ACLs and thus it is not really a read only or write only field but
something inbetween.

Previously the SerializerMethodField was used for the embargoed field
but that doesn't really goes well with generating the OpenAPI schema
as SerializerMethodField is read only and does not appear in request
body example. Instead we define a custom EmbargoedField with BooleanField
logic. On create or update the field is serialized the classic way,
ACLs are checked and then validation is aborted since we don't have
a field to write in so instead we modify ACLs based on embargoed value.