Fix embargoed str/bool conversion
This fixes serious problem with setting the ACLs. When we obtain the
embargoed from request body it is string not bool but we are working
with it like with bool and thus ACLs are not correctly set.

This was not capcured by the tests since in tests we supply the dict with
data and we are using python types (True/False bool values) which are
then directly passed from request body and thus values are already
in expected formats. Most basic solution is to just check whether the
embargoed is string and convert it to bool.