Add check for breaking changes in OpenAPI schema
This commit adds a GitHub workflow action that will run the oasdiff
tool on the base branch's openapi.yml and the head branch's
openapi.yml on every pull request, and if the tool detects any
backwards-incompatible changes the check will fail.

This commit also makes these checks no longer bound to PRs with a base
branch of `master`, as it can happen with `release-x.y.z` branches too.