Raise DataInconsistencyError on Bugzilla concurrent update error
Bugzilla does not seem to support concurrent updates at all, or if it
does they are extremely unpredictable to the point that they cannot be
relied on. This means that any concurrent calls to the OSIDB API that
might trigger a Bugzilla update can fail, previously this would yield a
simple HTTP 400 error that bubbles up from the python-bugzilla library
and it would be quite hard to figure out what the origin of said HTTP
400 error could be.

With this commit we will catch said errors and suggest to whomever
catches the error that the most likely reason for the error is a
concurrent update which Bugzilla does not support, and to try again
later.