Add meta_attr field to FlawReference model
In Bzimport we pass this attribute to the
FlawReferenceManager.create_flawreference() method which ends up passing
it to the FlawReference constructor, which will raise a TypeError as the
field does not exist in the model, this effectively makes all
synchronizations of new CVEs invalid.

This commit simply adds the attribute to the FlawReference model and
creates the relevant migrations for it.

Closes OSIDB-990