This PR introduces new flaw comment add API.
Closes OSIDB-81.

New comments are created as "pending", marked by empty
external_system_id. "Pending" comments are deleted on every
fetch from Bugzilla, this means that comments created through the OSIDB
API are not guaranteed to be referenceable since their UUID is likely to
change as these are OSIDB-only temporary comments.

However, this guarantees integrity in OSIDB's side because we will
always use BZ as canonical source of truth for comments, even if the
comments were originally created in OSIDB.

In a standard request-response lifecycle the following will happen when
creating a comment:

* Comment is created and stored in OSIDB without an external_system_id,
  this classifies the comment as a "pending" comment.
* The comment is posted to Bugzilla
* When we save the object, we force fetching from bugzilla, and the
  collector will delete all pending comments for a given flaw and
  recreates comments as necessary, this means that technically speaking
  there is a very narrow window (RTT) in which a "pending" comment
  exists.
* Technically speaking we have a chance of having duplicate comments on
  concurrent requests, but the window is so narrow that in practice it
  didn't happen during testing (in fact pending flaws were invisible in
  the database during testing).