Properly serialize instances of the Profile model
The __str__() method for the Profile model returns self.username which
is not an attribute defined in the Profile class itself, instead it is
defined in the `User` model one-to-one relationship.

This meant that trying to serialize a Profile object would result in an
AttributeError, we fix this by creating a Python property that simply
accesses self.user.username.