Include raw summary from BZ in Flaw.meta_attr
The Flaw.title attribute is based off of the `summary` attribute from
Bugzilla, however before setting Flaw.title the summary goes through
various transformations which might not be perfect and the resulting
title might be incorrect.

With this commit we include the raw BZ version of the title (summary in
BZ) so that clients have the full context should they need it.