Copy FlawMeta ACKNOWLEDGMENT into FlawAcknowledgment.
This makes the DB data consistent and ready for FlawAcknowledgment
bbsync. Without this consistency, FlawAcknowledgment manipulation
and bbsync would be either too complex or nonfunctional, because
it wouldn't be clear whether to assemble srtnotes acknowledgments
based on FlawMeta or FlawAcknowledgment, or would make acks disappear
if a Flaw without populated FlawAcknowledgment was bbsynced.