Vendor deprecate_field to allow for more dynamic return_instead values
With this commit we extend django-deprecate-fields and provide our own
version of the deprecate_field decorator, in this version the
return_instead parameter can take a callable that accepts either 0 (like
the original) or 1 argument.

If an argument can be passed, it shall be the Model object instance on
which the field access was requested. This is useful in cases where the
deprecation happens due to a rename or restructuring and can be
extrapolated from another Model object instance attribute.