Lower the maximum amount of records per page when querying BZ
1000 is technically the maximum amount of records that the Red Hat
bugzilla API will allow on each individual query, however if the query
takes longer than 120s to return the server will kill it and return a
503 error.

This can lead to collectors or anything that uses the BugzillaQuerier
class to fail and/or get stuck in an endless loop. With this commit we
make a choice to reduce the amount of records requested per page to an
order of magnitude lower: 100, this means that OSIDB might be slower
when performing a sync from scratch but more or less ensures that we
won't get any bugzilla collectors stuck.

Closes OSIDB-1232