Set AFFECTED as highest precedence for delegated_resolution
For Affects that have a resolution of DELEGATED, we look at its
trackers' own state/resolution to determine the actual delegated
resolution of the Affect.

If there's more than a single Tracker for an Affect, we take the
Tracker's state/resolution with highest precedence or "weight" and this
precedence is determined by a list of possible resolutions.

Before this commit, said list was ordered in such a way that the
NOTAFFECTED resolution had highest precedence however this is a mistake,
as if we have 5 trackers, 4 with NOTAFFECTED and 1 with AFFECTED
resolution we should determine that the Affect is AFFECTED, not
NOTAFFECTED.

This commit solves the issue by changing the order of the AFFECTED and
NOTAFFECTED resolutions in the aforementioned list.

Closes OSIDB-1230