Implement weighted replica database router
This (highly specific) database router distributes the load on the
database based on the nature of the request.

For read operations, a third of the requests will go to the default or
primary database and the remaining two thirds of the requests will go to
the read-only replica.

For any other operations, nothing changes.

This should reduce the load on the database and/or potentially allow for
more concurrent requests and generally improve performance.