Disarm potential SQL injection vector
Per Django's documentation [1] and generally any Python database
driver library (such as psycopg2) it is generally a bad idea to directly
do string interpolation when said interpolation consists of
user-provided parameters into an SQL statement.

Instead, users should:

* Avoid manually quoting identifiers within the SQL statement
* Pass any parameters as second argument of the execute/raw function

While in theory this represents a vector for SQL Injection, it is not
exploitable in practice because all input interpolated into the SQL
statement is converted to UUIDs so there is no risk.

[1] https://docs.djangoproject.com/en/3.2/topics/db/sql/#passing-parameters-into-raw