Apply ACLs to all database connections
Given the nature of the WeightedReplicaRouter, all write operations
(including the setting of database parameters) go to the primary
database even if we end up fetching from another database, in this case
the replica. This means that when fetching from the replica ACLs might
not be set and can lead to different results from one request to
another.

By applying ACLs to all related databases (in this case the default and
replica) we guarantee that fetching from either database is possible and
consistent.