Bypass row-level security when applying latest data migration
Because of the way that OSIDB handles access to records, database users
don't have access to any rows by default, this would lead to
ObjectDoesNotExist errors when querying certain records from certain
tables and would ultimately lead to the migration's failure.

To fix this, we simply set the user ACLs before executing the migration
and we grant ourselves every permission (r/w public/embargo).

This commit also updates the developer documentation so that this step
is a bit more clear.