Enable better logging of celery and bugzilla
This commit makes the necessary changes to the various settings files to
allow for better, more granular logging of the python-bugzilla library
and of the celery workers.

In particular:

* The celery logging handler will now accept logs from any level
* Celery pods will now log anything at INFO level or above
* python-bugzilla will now log anything at DEBUG level or above

The logging of python-bugzilla should only happen in environments where
`DEBUG=True`.

This is done in order to be able to have a better record of what is
being sent/requested from the Red Hat Bugzilla instance.