Use proper locking algorithm for Collector tasks
This commit changes the makeshift locking algorithm that was originally
implemented as part of the Collector framework to the Redlock algorithm
using the rhubarb library.

Since the rhubarb library is still early into development and PyPI has
temporarily disabled new registrations, its installation is done via
direct wheel from GitHub.

This commit removes the various reset and similar mechanisms as they
only worked under certain very specific conditions, previously it was
possible for a worker to die and for the collector state to remain in
RUNNING for a really long time, effectively creating a deadlock.

The commit also makes some modifications to the existing collector
dependency behavior, basically instead of automatic retries for blocked
collector tasks we simply ignore said tasks, this is done not only to
prevent an unintended lock release from rhubarb but also because it can
clog up the task queue unnecessarily.

Closes OSIDB-1521