Introduce usage of redbeat celery task scheduler
This commit adds celery-redbeat as a dependency of the project and sets
it as the default scheduler for celery tasks.

Unlike celery-beat, redbeat uses a Redis instance for keeping track of
task schedules and uses a distributed mutex lock in order to ensure that
only one instance of the scheduler can actually schedule tasks.

This comes with the following benefits:

* The scheduler can have high availability without compromising
  scheduling, meaning that we can have e.g. 3 pods of the scheduler
  running but only one will be actively doing work at any given time, if
  two Availability Zones happen to go down, the remaining pod will pick
  up the slack.
* The schedule is stored in Redis, meaning that even if there are
  scheduler restarts, the work to be done can be picked up immediately
  without delays.
* Tasks can be dynamically created/modified.
* Faster startup.

This commit also simplifies the way that the celery-beat scripts work.

Closes OSIDB-1524