Fix OpenLDAP test data file
Starting a new OpenLDAP container service has been failing since the
introduction of the internal groups.

The reason being that OpenLDAP complains and shutdowns if the imported
data contains empty groups (i.e. no members), this was not visible
unless running non-automated integration tests so it was easily missed.

This commit also modifies the test docker-compose file so that the
OpenLDAP service container does not depend on the osidb-data PG
database, as it doesn't really depend on it.

The commit also introduces a debugging env variable to the OpenLDAP
container, as without it it's pretty hard to know what went wrong when
spinning up the OpenLDAP container.