Defer evaluation of auth_client fixture
This fixture is problematic when used in combination with the enable_rls
autouse fixture due to the order in which these fixtures are called but
also because of the way that RLS/ACLs work in OSIDB.

When running any test, the `enable_rls` fixture is automatically applied
which essentially sets the ACLs at the database connection level so that
they can be bypassed in tests that do not explicitly test RLS behavior.

When running any test that requests the `auth_client` fixture, first the
`enable_rls` autouse fixture is applied, but then that work is undone by
the `auth_client` fixture which sets the ACLs to the user that logged in
using the client, and this happens even if the auth_client is not used
until later on in the test, after setting up all the test data.

By deferring the login action we prevent the undoing of the ACLs bypass,
this already fixes about ~50% of tests that were previously failing
because of this conflict and is a cheap change to make.