Refactor ACL tests and fixtures to use user-based ACLs
This commit gets rid of the embargo_access fixture which granted the
default testuser extra ACLs and replaces it with a mechanism in the
auth_client fixture which lets users of the fixture select an LDAP user
to login as, each different LDAP user having a unique set of ACLs.

This makes it easier to test ACLs as there is no need to manually fiddle
with Django groups, different users can have different access levels and
more can be added if needed.

This commit also makes it so that for local and ci environments, the
active LDAP group is what determines authentication access to OSIDB
instead of the data-prodsec, etc. groups, as these groups explicitly
grant access to read public data, the reason behind this is to test
users that can authenticate but have no authorization to read any data.