Rename test models to remove pytest warnings
Any class that starts with `Test` that is found in pytest's path is
treated as a test and is collected by pytest as a test if possible.

In practice it doesn't happen for these classes because they implement
an __init__ method which pytest considers as "not a test", but it's good
practice to simply not introduce model classes with a name starting with
`Test`.