Improve performance of changed_{before,after} filter
This commit achieves a jump in performance of up to 28% for queries that
use the changed_{before,after} filters by using database
denormalization.

Before this commit, the aforementioned filters would work by extending
the queryset to filter flaws, affects and trackers that had been updated
in relation to the datetime passed in as query parameter, this is slow
because it will perform multiple joins of very big tables and then
ultimately slap a big ol' DISTINCT on top of it to make matters worse.

With this commit, a new local_updated_dt field is introduced in the Flaw
model with a corresponding index, this local_updated_dt field is updated
each time a Flaw, its affects or its trackers are updated, but since
it's kept in the main table (flaws) no joins nor distinct are necessary,
so in the best-case scenario (e.g. getting CVE ids of flaws that changed
in some timeframe) the response time improves by up to 28%.

In the worst-case scenario (e.g. requesting all data, performing the
joins anyway) we can observe up to a 10% performance increase.