Set up make entrypoint to launch web-server with gunicorn
The default Django server is not meant for production use as it is not
ready to handle multiple concurrent requests.

By setting up an alternative entrypoint that launches the service using
the gunicorn WSGI server, it is easier to reproduce a stage/production
environment's ability to handle requests, this is specially useful when
doing performance testing.