Move auto-adjust logic from constructor to pre_save signal
Per the Django docs [1], it is seldom a good idea to introduce custom
behavior during a model's `__init__()` method, especially if such an
override would end up accessing model fields, as this can lead to
recursion errors in such cases.

One particular example is the use of `only()` in a queryset, because of
the `__init__` override in `WorkflowModel`, calling `only()` on the
`Flaw` model would yield a RecursionError.

With this commit, the behavior is moved to the `pre_save` signal, as we
don't really need this behavior until we're saving to the database.

[1] https://docs.djangoproject.com/en/5.0/ref/models/instances/#creating-objects