Dynamically fetch/prefetch based on include_fields filter content
This commit performs some optimizations on the include_fields filter,
instead of fetching everything from the database for a given table and
then filtering the output in Python during serialization, we change the
queryset based on user-input to only prefetch the relevant relations and
only fetch the relevant fields.

This leads to drastic boosts in performance (of up to an order of
magnitude) which are sorely needed especially for the /flaws endpoint.