Handle both Affect and Tracker objects during m2m_changed signal
Prior to this commit, the fact that the instance parameter in an
m2m_changed signal can be an instance of either of the two ends of the
m2m relationship was not taken into account.

This meant that the m2m_changed signal would only expect objects of one
type of the relationship, and if an object from the other side of the
relationship was passed through it would fail as the attributes being
checked are not the same.

With this commit, we properly handle both ends of the relationship.

Closes OSIDB-2059