Ignore invalid fields in include_fields filter
Prior to this commit and introduced by the recent performance
optimizations done to the include_fields filter, it was possible to make
the server return a 500 response whenever the following conditions were
met:

1. At least one valid record exists in the database for the queried
   resource.
2. An invalid field containing the double underscore separator (__) was
   passed to the include_fields filter.

This is due to the heuristic determining the fields to prefetch based on
the aforementioned separator and there being a presumed valid field left
of the separator.

However, as we've learned, queries that requested non-existing fields or
non-sensical fields such as __placeholder_field which prominently do not
have a valid field left of the separator, are real-world use-cases of
the API and break the heuristic as the field to prefetch suddenly
becomes the empty string ("").

This commit fixes this issue in order to guarantee
backwards-compatibility by ensuring that there's at least a valid field
left of the separator, if there isn't we simply ignore it and skip the
optimization for that given field (since it's not a real field).

The overall stability of the heuristic is more solid but also more
concise now.

Closes OSIDB-2048