Speed up exclude_fields filter by defering unwanted fields
By automatically deferring the fields passed to the exclude_fields
filter, a performance optimization of about ~15% or 0.2 RPS can be
achieved.

Unfortunately determining the relationships to prefetch in the case of
exclude_fields is a lot more complicated than for include_fields so this
optimization only handles top-level or model-level attributes.

Closes OSIDB-1871