Update valid affectedness-resolution combinations
Some affectedness-resolution combinations have been removed to comply
with the combinations defined in SFM2.

However, existing data is not to be modified as old combinations may
still be pulled from a collector without running the validation, so the
original enumerations have not been changed.

If an affect is created which uses an historical combination, an alert
will be raised.