Set migrated/duplicated delegated resolution to not affected
Before this commit, the MIGRATED and DUPLICATE resolutions which can
come from BZ were not contemplated in osidb, and in these cases, they
get assigned a state of AFFECTED by default. Since AFFECTED has the
highest precedence, even if there is another tracker that says it's
not a bug, the delegated resolution is AFFECTED which is wrong.

With this commit, trackers with MIGRATED/DUPLICATE resolutions are set
a NOTAFFECTED state, since it is assumed that if there is such a
tracker, then there is at least another one with actual information
about the state, and since NOTAFFECTED has the lowest precedence, the
final delegated resolution will be that of the other tracker with a
valid state.

Closes OSIDB-1406