Make ps_product property available in affect API
The `ps_product` is not an attribute of Affect itself, but rather an
attribute of its `ps_module`. With this commit, we make the `ps_product`
property directly available from the affect API by linking to its
module's product. This is useful e.g. for grouping affects by their
PS product names instead of the modules.