Make workflows API RESTful
The workflows API does not follow REST priciples. Right now, the way to
access a workflow is
${SERVICE_URL}/workflows/v1/api/<workflow_uuid>
but for it to be RESTful, the name of the resource being requested
should go before the ID, so it should look like
${SERVICE_URL}/workflows/v1/api/workflows/<workflow_uuid>

This was the case before the endpoint was renamed from osim to
workflows, and the osim endpoint deprecated.
With this commit, the URL will behave as before, making it comply with
REST principles again.

The previous endpoints, as well as the deprecated osim ones, have been
removed. This breaking change is acceptable since the workflows API is
still experimental and not being used in production, so it is not
necessary to keep the old endpoints.