Allow editing flaws in NEW workflow state
There is a validation in flaws which enforces them to have at least one
affect, which is only disabled when creating a flaw. This commit also
disables this validation when editing 'draft' flaws, i.e. flaws which
are in a NEW state, and instead issues an alert. The validation is still
active in other states and when trying to move to the next state.

Closes OSIDB-2452.